Adaptação do projeto

+++ src/main/java/br/com/telematica/siloapi/repository/TipoSiloRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.telematica.siloapi.model.entity.TipoSiloEntity;
 
-public interface TipoSiloRepository extends JpaRepository<TipoSiloEntity, Integer> {
+public interface TipoSiloRepository extends JpaRepository<TipoSiloEntity, Long> {
 
 	void deleteByTsicod(Integer tsicod);
